Restructure biome::key_management module
Refactors the store, database, and postgres modules to all
be under the key_management::store module. This places the
database code specific to diesel under the store::diesel
module which allows for other store backends. This also
moves all PostgreSQL-specific code to diesel::postgres
which allows for other database backend implementations.

Also moves helpers to diesel::operations.
