+++ cli/src/action/database.rs
@@ -20,7 +20,7 @@ use diesel::{connection::Connection as _, pg::PgConnection};
 #[cfg(feature = "database-migrate-biome-credentials")]
 use splinter::biome::credentials::store::run_postgres_migrations as run_biome_credentials_migrations;
 #[cfg(feature = "database-migrate-biome-key-management")]
-use splinter::biome::key_management::database::postgres::run_migrations as run_biome_key_management_migrations;
+use splinter::biome::key_management::store::run_postgres_migrations as run_biome_key_management_migrations;
 #[cfg(feature = "database-migrate-biome-notifications")]
 use splinter::biome::notifications::store::run_postgres_migrations as run_biome_notifications_migrations;
 #[cfg(feature = "database-migrate-biome-user")]
