+++ libsplinter/src/biome/key_management/mod.rs
@@ -14,11 +14,9 @@
 
 //! Provides an API for storing key pairs and associating them with users.
 
-pub mod database;
+pub mod store;
 
-pub(in crate::biome) mod store;
-
-use database::postgres::models::KeyModel;
+use store::diesel::models::KeyModel;
 
 // Represents a public and private key pair
 #[derive(Serialize)]
