+++ libsplinter/src/biome/key_management/store/diesel/operations/mod.rs
@@ -12,6 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod keys;
-
-pub use keys::{insert_key, list_keys, list_keys_with_user_id, update_key};
+pub(super) mod keys;
